Archived existing code
The first push in this repository. There are a lot more older pushes in the private Wonders of Patience Repository, as they were once in the same project folder.

I made this Github as a means of backing up files and archival, even if I am the one person working on the Gamemaker Project currently.